+++ public/app/features/alerting/unified/mocks/alertmanagerApi.ts
@@ -1,9 +1,12 @@
 import { http, HttpResponse } from 'msw';
 import { SetupServer } from 'msw/node';
 
+import { mockAlertmanagerAlert } from 'app/features/alerting/unified/mocks';
+
 import {
   AlertmanagerChoice,
   AlertManagerCortexConfig,
+  AlertState,
   ExternalAlertmanagersResponse,
 } from '../../../../plugins/datasource/alertmanager/types';
 import { AlertmanagersChoiceResponse } from '../api/alertmanagerApi';
@@ -13,8 +16,12 @@ export const defaultAlertmanagerChoiceResponse: AlertmanagersChoiceResponse = {
   alertmanagersChoice: AlertmanagerChoice.Internal,
   numExternalAlertmanagers: 0,
 };
+
+export const alertmanagerChoiceHandler = (response = defaultAlertmanagerChoiceResponse) =>
+  http.get('/api/v1/ngalert', () => HttpResponse.json(response));
+
 export function mockAlertmanagerChoiceResponse(server: SetupServer, response: AlertmanagersChoiceResponse) {
-  server.use(http.get('/api/v1/ngalert', () => HttpResponse.json(response)));
+  server.use(alertmanagerChoiceHandler(response));
 }
 
 export const emptyExternalAlertmanagersResponse: ExternalAlertmanagersResponse = {
@@ -38,3 +45,17 @@ export function mockAlertmanagerConfigResponse(
     )
   );
 }
+
+export const alertmanagerAlertsListHandler = () =>
+  http.get('/api/alertmanager/:datasourceUid/api/v2/alerts', () =>
+    HttpResponse.json([
+      mockAlertmanagerAlert({
+        labels: { foo: 'bar', buzz: 'bazz' },
+        status: { state: AlertState.Suppressed, silencedBy: ['12345'], inhibitedBy: [] },
+      }),
+      mockAlertmanagerAlert({
+        labels: { foo: 'bar', buzz: 'bazz' },
+        status: { state: AlertState.Suppressed, silencedBy: ['12345'], inhibitedBy: [] },
+      }),
+    ])
+  );
