+++ public/app/features/alerting/unified/mocks/server/configure.ts
@@ -1,5 +1,5 @@
 import server from 'app/features/alerting/unified/mockApi';
-import { alertmanagerChoiceHandler } from 'app/features/alerting/unified/mocks/server/handlers';
+import { alertmanagerChoiceHandler } from 'app/features/alerting/unified/mocks/alertmanagerApi';
 import { AlertmanagerChoice } from 'app/plugins/datasource/alertmanager/types';
 
 /**
